Move DataGateConfigurationManager class to separate module
